:bug: accelerator L2 wrapper holds flush until last request is accepted
This solves a potential race condition: if the last line is written to
the accelerator's cache after the flush, a processor, or another
accelerator may request that line from the LLC before the accelerator's
GetM reaches the LLC.
Accesses are not protected by locks, because we always assume
accelerators execute in mutual exclusion on their dedicated memory
region. As long as flush is triggered after the last write has occured,
any further request to the LLC will be handled properly and may incur a
recall.